+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Update landing page and remove getting started button.
+
 ## [0.7.0] - 2023-09-15
 
 ### Added